docs: implement perceptron for iris dataset

+++ examples/linear_model/perceptron.py
@@ -36,13 +36,17 @@ X = df.iloc[0:100, [0, 2]].values
 
 # plot data
 plt.scatter(X[:50, 0], X[:50, 1], color="red", marker="o", label="setosa")
-plt.scatter(X[50:100, 0], X[50:100, 1], color="blue", marker="x", label="versicolor")
+plt.scatter(X[50:100, 0],
+            X[50:100, 1],
+            color="blue",
+            marker="x",
+            label="versicolor")
 plt.xlabel("sepal length [cm]")
 plt.ylabel("petal length [cm]")
 plt.legend(loc="upper left")
 plt.show()
 
-ppn = Perceptron(alpha=0.1, n_iters=10)
+ppn = Perceptron(alpha=np.float16(0.1), n_iters=np.int64(10))
 ppn.fit(X, y)
 
 plt.plot(range(1, len(ppn.errors_) + 1), ppn.errors_, marker="o")
